requirements.txt
anthropic>=0.18.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
playsound>=1.3.0
